backend/requirements.txt
annoy
fastapi
uvicorn
deepface
faiss-cpu
retina-face
ffmpeg-python
python-dotenv
python-multipart